+++ libs/core/src/handlers/events.rs
@@ -78,8 +78,6 @@ slu_events_declaration! {
 
     ColorsChanged(UIColors) as "colors-changed",
 
-    ToolbarOverlaped(bool) as "set-auto-hide",
-
     WegOverlaped(bool) as "set-auto-hide",
 
     WMSetReservation as "set-reservation",