+++ src/ui/react/toolbar/app.tsx
@@ -2,18 +2,25 @@ import { $system_colors } from "@shared/signals";
 import { useDarkMode } from "@shared/styles";
 import { getCurrentWebviewWindow } from "@tauri-apps/api/webviewWindow";
 import { ConfigProvider, theme } from "antd";
-import { useEffect } from "react";
 
 import { ErrorBoundary } from "../weg/components/Error/index.tsx";
 import { ErrorFallback } from "./components/Error/index.tsx";
 import { FancyToolbar } from "./modules/main/Toolbar.tsx";
+import { useSignalEffect } from "@preact/signals";
+import { $lastFocusedOnMonitor } from "./modules/shared/state/windows.ts";
 
+const webview = getCurrentWebviewWindow();
 export function App() {
   const isDarkMode = useDarkMode();
 
-  useEffect(() => {
-    getCurrentWebviewWindow().show();
-  }, []);
+  useSignalEffect(() => {
+    const fullscreened = !!$lastFocusedOnMonitor.value?.isFullscreened;
+    if (fullscreened) {
+      webview.hide();
+    } else {
+      webview.show();
+    }
+  });
 
   return (
     <ConfigProvider