+++ src/ui/react/toolbar/modules/item/infra/infra.tsx
@@ -4,14 +4,12 @@ import { useWindowFocusChange } from "@shared/hooks";
 import { Menu } from "antd";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useSelector } from "react-redux";
 
 import { BackgroundByLayersV2 } from "@shared/components/BackgroundByLayers/infra";
 
-import { Selectors } from "../../shared/store/app.ts";
-
 import { CommonItemContextMenu } from "./ContextMenu.tsx";
 import { InnerItem, type InnerItemProps } from "./Inner.tsx";
+import { $focused } from "../../shared/state/windows.ts";
 
 export function Item(props: InnerItemProps) {
   const [openContextMenu, setOpenContextMenu] = useState(false);
@@ -48,11 +46,7 @@ export function Item(props: InnerItemProps) {
 }
 
 export function AppsItem({ module, ...rest }: { module: GenericToolbarItem }) {
-  const window = useSelector(Selectors.focused) || {
-    name: "None",
-    title: "No Window Focused",
-    exe: null,
-  };
+  const window = $focused.value;
   return <Item {...rest} module={module} extraVars={{ window }} />;
 }