+++ src/ui/react/toolbar/modules/shared/store/domain.ts
@@ -4,14 +4,12 @@ import type {
   Battery,
   BluetoothDevice,
   File,
-  FocusedApp,
   MediaDevice,
   MediaPlayer,
   NetworkAdapter,
   PowerMode,
   PowerStatus,
   User,
-  WegAppGroupItem,
   WlanBssEntry,
 } from "@seelen-ui/lib/types";
 
@@ -25,7 +23,6 @@ export interface RootState {
   userPicturesFolder: File[];
   userVideosFolder: File[];
   userMusicFolder: File[];
-  focused: FocusedApp | null;
   env: Record<string, string>;
   bluetoothDevices: BluetoothDevice[];
   discoveredBluetoothDevices: BluetoothDevice[];
@@ -41,9 +38,4 @@ export interface RootState {
   mediaInputs: MediaDevice[];
   notifications: AppNotification[];
   languages: SystemLanguage[];
-  openApps: WegAppGroupItem[];
-  windowColorByHandle: Record<
-    string,
-    { background: string; foreground: string }
-  >;
 }